requirements_torch.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
absl-py=0.7.1=pypi_0
audioread=2.1.6=py37_0
backcall=0.2.0=pyhd3eb1b0_0
blas=1.0=mkl
bounded-pool-executor=0.0.3=pypi_0
bzip2=1.0.8=h7b6447c_0
ca-certificates=2021.1.19=h06a4308_0
certifi=2020.12.5=py37h06a4308_0
chardet=3.0.4=pypi_0
cudatoolkit=10.0.130=0
cycler=0.10.0=py37_0
dbus=1.13.18=hb2f20db_0
decorator=4.4.2=pyhd3eb1b0_0
expat=2.2.10=he6710b0_2
ffmpeg=4.2.2=h20bf706_0
fontconfig=2.13.0=h9420a91_0
freetype=2.10.4=h5ab3b9f_0
future=0.17.1=pypi_0
glib=2.66.1=h92f7085_0
gmp=6.1.2=h6c8ec71_1
gnutls=3.6.5=h71b1129_1002
grpcio=1.23.0=pypi_0
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
h5py=2.8.0=py37h3010b51_1003
hdf5=1.10.2=hba1933b_1
icu=58.2=he6710b0_3
idna=2.10=pypi_0
intel-openmp=2020.2=254
ipython=7.5.0=py37h39e3cac_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
jedi=0.17.0=py37_0
joblib=1.0.0=pyhd3eb1b0_0
jpeg=9b=h024ee3a_2
kiwisolver=1.3.0=py37h2531618_0
lame=3.100=h7b6447c_0
ld_impl_linux-64=2.33.1=h53a641e_7
libedit=3.1.20191231=h14c3975_1
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libllvm10=10.0.1=hbcb73fb_5
libopus=1.3.1=h7b6447c_0
libpng=1.6.37=hbc83047_0
librosa=0.6.3=py_0
libstdcxx-ng=9.1.0=hdf63c60_0
libuuid=1.0.3=h1bed415_2
libvpx=1.7.0=h439df22_0
libxcb=1.14=h7b6447c_0
libxml2=2.9.10=hb55368b_3
llvmlite=0.34.0=py37h269e1b5_4
markdown=3.1.1=pypi_0
matplotlib=3.1.0=py37h5429711_0
mir-eval=0.5=pypi_0
mkl=2020.2=256
mkl-service=2.3.0=py37he8ac12f_0
mkl_fft=1.2.0=py37h23d657b_0
mkl_random=1.1.1=py37h0573a6f_0
mxnet-cu102=1.6.0.post0=pypi_0
ncurses=6.2=he6710b0_1
nettle=3.4.1=hbb512f6_0
numba=0.51.2=py37h04863e7_1
numpy=1.19.2=py37h54aff64_0
numpy-base=1.19.2=py37hfa32c7d_0
opencv-python=4.5.1.48=pypi_0
openh264=2.1.0=hd408876_0
openssl=1.1.1i=h27cfd23_0
pandas=0.25.1=pypi_0
parso=0.8.1=pyhd3eb1b0_0
pcre=8.44=he6710b0_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.1.0=pypi_0
pip=20.3.3=py37h06a4308_0
pqdm=0.1.0=pypi_0
prompt_toolkit=2.0.10=py_0
protobuf=3.9.1=pypi_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pygments=2.7.4=pyhd3eb1b0_0
pyparsing=2.4.7=pyhd3eb1b0_0
pyqt=5.9.2=py37h05f1152_2
python=3.7.9=h7579374_0
python-dateutil=2.8.1=pyhd3eb1b0_0
python-graphviz=0.8.4=pypi_0
pytz=2021.1=pyhd3eb1b0_0
qt=5.9.7=h5867ecd_1
readline=8.1=h27cfd23_0
requests=2.24.0=pypi_0
resampy=0.2.1=py_1
scikit-learn=0.21.1=py37hd81dba3_0
scipy=1.5.2=py37h0b6359f_0
seaborn=0.9.0=pypi_0
setuptools=52.0.0=py37h06a4308_0
sip=4.19.8=py37hf484d3e_0
six=1.15.0=py37h06a4308_0
sounddevice=0.3.13=pypi_0
soundfile=0.10.2=pypi_0
sqlite=3.33.0=h62c20be_0
tabulate=0.8.7=pypi_0
tbb=2020.3=hfd86e86_0
tensorboard=1.14.0=pypi_0
tk=8.6.10=hbc83047_0
torch=1.2.0=pypi_0
torchsummary=1.5.1=pypi_0
torchvision=0.4.0=pypi_0
tornado=6.1=py37h27cfd23_0
tqdm=4.49.0=pypi_0
traitlets=5.0.5=pyhd3eb1b0_0
typing-extensions=3.7.4.3=pypi_0
urllib3=1.25.11=pypi_0
wcwidth=0.2.5=py_0
werkzeug=0.15.5=pypi_0
wheel=0.36.2=pyhd3eb1b0_0
x264=1!157.20191217=h7b6447c_0
xz=5.2.5=h7b6447c_0
zlib=1.2.11=h7b6447c_3
